Change how calls are added to the database
Stricter control on how calls are added. RBQFetchResultsController
provides us a way to see inserts of new objects. Prior to this, calls were
added AND THEN updated to their respective status. Thus, we were
unable to check the status of the call when it is first inserted.

This fix makes sure that the call has a status prior to being inserted
into the database.