Add default values for the call state when a call is incoming
Since audio state callbacks are only received when a friend sends a call control,
audio state is not known during incoming call. Thus if an incoming call is answered,
the state will be the default of 0 (paused state). This commit will give default values
to the call state for that event.